Added functionality to Options.vue and implemented config file
If config file does not exist, it is created in the root directory of the built app.
Options menu can now update the game folder and the default preset. Still trying to figure out how to get elements in other Vues to update when the user applies the new configuration.